Organized code + report files organizations
Cleaned report code, moved out from main into own module.
Now report files are created in a dedicated folder with name specified by argument (or default 'report') plus the datetime of the sniffing.
Folder has 777 priviledges to be easly handled by all users (since the process needs to be executed as root, files and dir would be created as root).
